refactor(peagen): use unified Task import
