+++ pkgs/standards/peagen/peagen/cli/commands/templates.py
@@ -9,7 +9,7 @@ import httpx
 import typer
 
 from peagen.handlers.templates_handler import templates_handler
-from peagen.models import Task
+from peagen.models.tasks import Task
 
 # ──────────────────────────────────────
 DEFAULT_GATEWAY = "http://localhost:8000/rpc"
