+++ pkgs/standards/peagen/peagen/handlers/control_handler.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from typing import Iterable
 
 from peagen.plugins.queues import QueueBase
-from peagen.models import Task
+from peagen.models.tasks import Task
 from peagen.core import control_core
 from peagen import defaults
