+++ pkgs/standards/peagen/peagen/handlers/doe_process_handler.py
@@ -10,7 +10,8 @@ import uuid
 import yaml
 
 from peagen.core.doe_core import generate_payload
-from peagen.models import Task, Status
+from peagen.models.tasks import Task
+from peagen.models import Status
 from peagen._utils.config_loader import resolve_cfg
 from peagen.plugins import PluginManager
 from peagen.plugins.storage_adapters.file_storage_adapter import FileStorageAdapter
