+++ pkgs/standards/peagen/peagen/handlers/extras_handler.py
@@ -8,7 +8,7 @@ from typing import Any, Dict
 from peagen._utils import maybe_clone_repo
 
 from peagen.core.extras_core import generate_schemas
-from peagen.models import Task
+from peagen.models.tasks import Task
 from .repo_utils import fetch_repo, cleanup_repo
