+++ pkgs/standards/peagen/peagen/handlers/fetch_handler.py
@@ -13,7 +13,7 @@ from pathlib import Path
 from typing import Any, Dict, List
 
 from peagen.core.fetch_core import fetch_many
-from peagen.models import Task  # for type hints only
+from peagen.models.tasks import Task  # for type hints only
 
 
 async def fetch_handler(task_or_dict: Dict[str, Any] | Task) -> Dict[str, Any]:
