+++ pkgs/standards/peagen/peagen/handlers/validate_handler.py
@@ -7,7 +7,7 @@ from typing import Any, Dict
 from peagen._utils import maybe_clone_repo
 
 from peagen.core.validate_core import validate_artifact
-from peagen.models import Task
+from peagen.models.tasks import Task
 
 
 async def validate_handler(task_or_dict: Dict[str, Any] | Task) -> Dict[str, Any]:
