+++ pkgs/standards/peagen/peagen/models/tasks/__init__.py
@@ -1,0 +1,3 @@
+from .task import Task
+
+__all__ = ["Task"]
